requirements.txt
discord.py
google-genai
httpx
python-dotenv
groq
pytest
pytest-asyncio
